+++ trunk/src/jpeg/jdinput.c
@@ -46,8 +46,8 @@ initial_setup (j_decompress_ptr cinfo)
 	jpeg_component_info *compptr;
 
 	/* Make sure image isn't bigger than I can handle */
-	if ((long) cinfo->image_height > (long) JPEG_MAX_DIMENSION ||
-		(long) cinfo->image_width > (long) JPEG_MAX_DIMENSION)
+	if ((s32) cinfo->image_height > (s32) JPEG_MAX_DIMENSION ||
+		(s32) cinfo->image_width > (s32) JPEG_MAX_DIMENSION)
 		ERREXIT1(cinfo, JERR_IMAGE_TOO_BIG, (unsigned int) JPEG_MAX_DIMENSION);
 
 	/* For now, precision must match compiled-in value... */
@@ -84,22 +84,22 @@ initial_setup (j_decompress_ptr cinfo)
 		compptr->DCT_scaled_size = DCTSIZE;
 		/* Size in DCT blocks */
 		compptr->width_in_blocks = (JDIMENSION)
-			jdiv_round_up((long) cinfo->image_width * (long) compptr->h_samp_factor,
-			(long) (cinfo->max_h_samp_factor * DCTSIZE));
+			jdiv_round_up((s32) cinfo->image_width * (s32) compptr->h_samp_factor,
+			(s32) (cinfo->max_h_samp_factor * DCTSIZE));
 		compptr->height_in_blocks = (JDIMENSION)
-			jdiv_round_up((long) cinfo->image_height * (long) compptr->v_samp_factor,
-			(long) (cinfo->max_v_samp_factor * DCTSIZE));
+			jdiv_round_up((s32) cinfo->image_height * (s32) compptr->v_samp_factor,
+			(s32) (cinfo->max_v_samp_factor * DCTSIZE));
 		/* downsampled_width and downsampled_height will also be overridden by
 		 * jdmaster.c if we are doing full decompression.  The transcoder library
 		 * doesn't use these values, but the calling application might.
 		 */
 		/* Size in samples */
 		compptr->downsampled_width = (JDIMENSION)
-			jdiv_round_up((long) cinfo->image_width * (long) compptr->h_samp_factor,
-			(long) cinfo->max_h_samp_factor);
+			jdiv_round_up((s32) cinfo->image_width * (s32) compptr->h_samp_factor,
+			(s32) cinfo->max_h_samp_factor);
 		compptr->downsampled_height = (JDIMENSION)
-			jdiv_round_up((long) cinfo->image_height * (long) compptr->v_samp_factor,
-			(long) cinfo->max_v_samp_factor);
+			jdiv_round_up((s32) cinfo->image_height * (s32) compptr->v_samp_factor,
+			(s32) cinfo->max_v_samp_factor);
 		/* Mark component needed, until color conversion says otherwise */
 		compptr->component_needed = TRUE;
 		/* Mark no quantization table yet saved for component */
@@ -107,8 +107,8 @@ initial_setup (j_decompress_ptr cinfo)
 	}
 	/* Compute number of fully interleaved MCU rows. */
 	cinfo->total_iMCU_rows = (JDIMENSION)
-		jdiv_round_up((long) cinfo->image_height,
-		  (long) (cinfo->max_v_samp_factor*DCTSIZE));
+		jdiv_round_up((s32) cinfo->image_height,
+		  (s32) (cinfo->max_v_samp_factor*DCTSIZE));
 	/* Decide whether file contains multiple scans */
 	if (cinfo->comps_in_scan < cinfo->num_components || cinfo->progressive_mode)
 		cinfo->inputctl->has_multiple_scans = TRUE;
@@ -157,11 +157,11 @@ initial_setup (j_decompress_ptr cinfo)
 
 		/* Overall image size in MCUs */
 		cinfo->MCUs_per_row = (JDIMENSION)
-			jdiv_round_up((long) cinfo->image_width,
-			(long) (cinfo->max_h_samp_factor*DCTSIZE));
+			jdiv_round_up((s32) cinfo->image_width,
+			(s32) (cinfo->max_h_samp_factor*DCTSIZE));
 		cinfo->MCU_rows_in_scan = (JDIMENSION)
-			jdiv_round_up((long) cinfo->image_height,
-			(long) (cinfo->max_v_samp_factor*DCTSIZE));
+			jdiv_round_up((s32) cinfo->image_height,
+			(s32) (cinfo->max_v_samp_factor*DCTSIZE));
 
 		cinfo->blocks_in_MCU = 0;
